requirements.txt
# USING GOOGLE GENERATIVE AI
# google-generativeai
# python-dotenv
# chromadb
# langchain-community
# rich

# USING OLLAMA
ollama
langchain-ollama
python-dotenv
chromadb
langchain-community
rich
langchain-text-splitters 